refactor(coin-mina): rename api folder to network
