+++ libs/coin-modules/coin-mina/src/logic/history/getBlockInfo.ts
@@ -1,4 +1,4 @@
-import { rosettaGetBlockInfo, RosettaBlockInfoResponse } from "../../api";
+import { rosettaGetBlockInfo, RosettaBlockInfoResponse } from "../../network";
 
 export const getBlockInfo = async (blockHeight: number): Promise<RosettaBlockInfoResponse> => {
   const data = await rosettaGetBlockInfo(blockHeight);
