+++ libs/coin-modules/coin-mina/src/logic/transaction/getFees.ts
@@ -1,5 +1,5 @@
 import BigNumber from "bignumber.js";
-import { fetchTransactionMetadata } from "../../api";
+import { fetchTransactionMetadata } from "../../network";
 import { Transaction } from "../../types/common";
 import { isValidAddress } from "../utils";
