+++ libs/coin-modules/coin-mina/src/network/types.ts
