+++ libs/coin-modules/coin-mina/src/types/index.ts
@@ -1,3 +1,3 @@
 export * from "./common";
 export * from "./signer";
-export type { ValidatorInfo } from "../api/types";
+export type { ValidatorInfo } from "../network/types";
